refactor: extract voice panel slot rendering and drag handlers
Extract useVoicePanelSlots hook (429→111 LOC) into focused modules:
- useVoicePanelDragHandlers: combined internal/external drag logic
- useVoicePanelSlotRendering: complex JSX rendering components

Achieves <300 LOC architectural target while maintaining full functionality
and backward compatibility through re-exports.
